pyproject.toml
[project]
name = "logical-data-modeling-assistant"
version = "1.0.0"
description = "A production-ready FastAPI backend that leverages LLMs to help users design, refine, and explain logical data models through a conversational interface."
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["fastapi", "llm", "data-modeling", "ai", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi>=0.112.4,<0.113",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "openai>=1.93.1",
    "anthropic>=0.57.1",
    "groq>=0.13.1",
    "instructor[groq]==1.9.2",
    "requests>=2.32.4",
    "httpx[http2]>=0.27.0",
    "tenacity>=9.1.2",
    "numpy>=2.3.1",
    "scikit-learn>=1.7.0",
    "fastmcp",
    "fastapi-mcp",
    "mcp[cli]>=1.12.1",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "uuid7>=0.1.0",
    "motor>=3.7.1",
    "python-jose[cryptography]>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/logical-data-modeling-assistant"
Repository = "https://github.com/yourusername/logical-data-modeling-assistant"
Documentation = "https://github.com/yourusername/logical-data-modeling-assistant#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "schemas"]
